refs #1279 [利润相关柱状图] * 完成柱状图的显示 * 完成 7,8 月份数据的测试

+++ core/app/models/view/post/SellingRecordChartsPost.java
@@ -202,7 +202,7 @@ public class SellingRecordChartsPost extends Post<HighChart> {
             @Override
             public void each(HighChart highChart, Date date, Map<String, Object> row) {
                 float profit = NumberUtils.toFloat(row.get("profit").toString());
-                float sales = NumberUtils.toInt(row.get("sales").toString());
+                float sales = NumberUtils.toFloat(row.get("sales").toString());
                 highChart.line("销售利润率").yAxis(1).add(date, (sales == 0) ? 0 : (profit / sales));
             }
         });
